feat(custom_watch): give guided_runs a glance page, widen the pages mask to u64
Page::GuidedRun renders the armed scripted coach run — target duration, the
cue the elapsed time has reached, the countdown to the next one, time left as
the hero — closing the live-run cluster after Pacer. The view is derived per
snapshot by running the ported cues_due over (-1, elapsed], so it holds no
per-tick dispatcher state and re-derives after a reboot mid-run; the setter
guards on library membership plus is_guided_run_valid and ignores an unknown
id rather than disarming an armed run. No cue text and no voice: the keys are
identifiers, the phone speaks the line.

Page 33 overflowed two ceilings. Page::bit() filled a u32 exactly, and in a
release build 1 << 32 wraps onto the Dashboard's bit, so the mask is now u64
end to end; the SET1 wire stays 32 bits and page::mask_from_wire leaves the
pages it cannot address enabled rather than silently curating them out. The
page grid's 4x8 body seats 32 cells, so grid_rows/grid_cell take the cursor
and the body scrolls in whole rows when the enabled set outgrows a screenful
(never at a real run's ~12-page mask).

+++ apps/custom_watch/core/src/input_flow.rs
@@ -130,7 +130,7 @@ pub fn grid_cursor_op(key: GridCursorKey, held_past_long: bool) -> GridCursorOp 
 }
 
 impl GridCursorOp {
-    pub fn apply(self, grid: &mut PageGrid, mask: u32) {
+    pub fn apply(self, grid: &mut PageGrid, mask: u64) {
         match self {
             GridCursorOp::Tap => grid.tap(mask),
             GridCursorOp::RowDown => grid.row_down(mask),
@@ -143,7 +143,7 @@ impl GridCursorOp {
 /// The page a BTN3 press lands on, walking the filtered cycle (data-present and
 /// curated) so a press never lands on an empty glance. An action that isn't a
 /// page walk leaves the page alone.
-pub fn paged(page: Page, action: Btn3Action, mask: u32) -> Page {
+pub fn paged(page: Page, action: Btn3Action, mask: u64) -> Page {
     match action {
         Btn3Action::PageNext => page.next_in(mask),
         Btn3Action::PagePrev => page.prev_in(mask),
@@ -321,7 +321,7 @@ mod tests {
 
     #[test]
     fn cursor_ops_drive_the_grid_and_the_two_directions_mirror() {
-        let mask = u32::MAX;
+        let mask = u64::MAX;
         let mut grid = PageGrid::open(Page::default(), mask);
         let opened_at = grid.cursor();
         grid_cursor_op(GridCursorKey::Forward, false).apply(&mut grid, mask);
@@ -338,7 +338,7 @@ mod tests {
 
     #[test]
     fn a_page_walk_follows_the_action_and_nothing_else_moves_the_page() {
-        let mask = u32::MAX;
+        let mask = u64::MAX;
         let page = Page::default();
         assert_eq!(paged(page, Btn3Action::PageNext, mask), page.next_in(mask));
         assert_eq!(paged(page, Btn3Action::PagePrev, mask), page.prev_in(mask));
@@ -353,7 +353,7 @@ mod tests {
 
     #[test]
     fn a_page_walk_wraps_the_cycle_in_both_directions() {
-        let mask = u32::MAX;
+        let mask = u64::MAX;
         let mut page = Page::default();
         let mut seen = 0usize;
         loop {
@@ -391,7 +391,7 @@ mod tests {
     fn the_page_walk_direction_matches_the_press_that_produced_it() {
         // The task derives the action from the press; the two must agree for
         // every run-view state, or a long press would page the wrong way.
-        let mask = u32::MAX;
+        let mask = u64::MAX;
         let page = Page::default();
         for state in [
             RecordState::Recording,
